Use sstableid and memtable ref to validate SAI ORDER BY
Vector uses a similarity score calculation to ensure
correctness of order. However, such computation is not
possible if do not have the source value stored in the
index or if we're unable to efficiently find the source
value. Instead, we can use the source value for a given
cell to determine if the index's source is the same.

When the materialized cell and the index's value
come from the same ss or memtable, we know that the
value was not subsequently updated and we therefore
know that the index's ordering is correct. If they
do not match, the ordering correctness is not known
(though it could still be correct if overwritten by
the same value), so we throw out the row to ensure
absolute correctness.

Future work will remove our reliance on scores and
ByteComparable objects and we'll replace them with
an overquery of at least one extra row per sstable
and memtable. This is required because switching
to validating source reference equality renders
index ordering results unable to be compared.

However, the one benefit is that we will not
need to do any computations or byte buffer
comparisons to determine order validity. Thus,
the new design will have some benefits and some
drawbacks. The largest benefit is supporting
a completely new class of types that are
eligible for ORDER BY clauses.