Remove isIndexDataValid reliance on vector sim score
There are several valid scenarios where the source
tables are not equal, but the rows could have scores
indicating the row is valid. See several tests
in VectorUpdateDeleteTest that fail on the last
commit. With this change, those tests are passing.
I include stricter assertions on row order to ensure
we do not have a subtle regression.

Note that the cases where the score would otherwise
indicate the row is valid are cases where we are
doing something suboptimal. However, these are
unlikely scenarios and not ones we need to optimize
for. Instead, we're prioritizing supporting more
generic order by.