Acquire and use consistent view for order by queries
Because we are using memtable references and sstable ids
to validate that a given cell in an ORDER BY result
out of an SAI index is valid, we need to make sure
that we have a single, consistent view of the indexes
and their associated mem/ss tables. The abstraction
could probably improve, but the underlying design
seems to work.

We do not use the lock on the boolean predicates because
those are validated using their actual predicate logic,
and therefore, we can skip any extra work to use the
ORDER BY view for the WHERE clauses.