Fix text encoding for hybrid queries
Enconding continues to be a challenge. In this
case, I found out that the inverted index gives
us a byte comparable based on the fixed byte
array for the entry. AFAICT, these bytes are
the bytes from the entry in storage. As such,
we needed to match the logic in the hybrid
stack to ensure correct validation of
byte comparable objects.