Test encoding everything with TypeUtil.asComparableBytes
This might not work, but I want to see what the
test results are. It might simplify things to ensure
that we always encode using the comparable bytes.

However, note that we are using the TypeUtil, which
is an SAI class and it encodes Inet, Varint, and Decimal
in a different way.