Improve method names on the Expression class
There is a lot of nuance here. I am not sure
how to fix the legacy complexity of how we
encode terms as we put them in the trie.
As such, I am pushing the ambiguity into the
method names to make it clear that there is
some form of encoding, but it's not actually
the full form.