CNDB-10907: Implement overlap diagnostics (#1453)
Provide a new --overlap (-O) option to nodetool compactionstats that generates an overlap map showing the maximum overlaps for every shard of every level in the strategy, as well as a total for all sstables in the system (not necessarily the sum of the per-level overlaps).

Change nodetool compactionstats to accept a keyspace and list of tables to restrict the --aggregate and --overlap output.